LayoutTests/platform/win/accessibility/list-role.html

<html>
    <head>
        <script src="../../../fast/js/resources/js-test-pre.js"></script>

        <script>
            if (window.testRunner)
                testRunner.dumpAsText();
        </script>
    </head>
    <body id="body">
        <p>
            This test verifies that, on Windows, lists return the "list" role.
            <br><br>
            <a href="https://bugs.webkit.org/show_bug.cgi?id=32736">Bug 32736 - REGRESSION(r52233): MSAA: Accessibility role of lists is wrong</a>
        </p>

        <ul id="list" tabindex=1>
            <li>list item</li>
        </ul>

        <p id="console"></p>

        <script>
            if (window.accessibilityController) {
                document.getElementById("list").focus();

                var list = accessibilityController.focusedElement;
                shouldBe('list.role', '"list"');
            }
        </script>
    </body>
</html>
